move helpers to helpers and create new lib file
