+++ tests/analytics.rs
